added support for css parts
I actually introduced a multi-config and dev server in order to make
sure everything works.

This introduced a lot of complexity, next step should be another
round of simpliciation. I might be forced to use strings, maybe
something like: `other-loader!part-loader!resource-if-available`
(if optional and present, with array never).
And use `loaderContext` to communicate part information.
Anyway: food for thought.